servers/requirements.txt
# APEX MCP Server — Pinned Dependencies
# Do NOT upgrade mcp without testing — breaking changes are common
fastmcp
notion-client>=2.2.1
httpx>=0.27.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
